"Working" with naive velocity calculation.
The velocity should be the same every day, if work is completed at that speed
every day. At least, the variance between days should be minimized.

It's working at least, but it telling me I need to do ~4 points per day.
That's a small project or week of small homework. Yeesh.